Code cleanup and speedup
also make compilation more stricter

+++ src/engine/translations.cpp
@@ -53,7 +53,7 @@ u32 crc32b( const char * msg )
         crc ^= static_cast<u32>( msg[index] );
 
         for ( int bit = 0; bit < 8; ++bit ) {
-            size_t mask = ~( crc & 1 );
+            uint32_t mask = ~( crc & 1 );
             crc = ( crc >> 1 ) ^ ( 0xEDB88320 & mask );
         }
