+++ src/fheroes2/dialog/dialog_armyinfo.cpp
@@ -469,7 +469,7 @@ void DrawBattleStats( const fheroes2::Point & dst, const Troop & b )
         if ( widthDiff > 0 ) {
             if ( space > 10 )
                 space = 10;
-            ow = dst.x + ( spritesWidth + space * ( spellsInfo.size() - 1 ) ) / 2;
+            ow = dst.x + ( spritesWidth + space * static_cast<int>( spellsInfo.size() - 1 ) ) / 2;
         }
         else {
             ow = dst.x + maxSpritesWidth / 2;
